expose ./core to tree-shake out the css-analyzer
