+++ .github/workflows/release.yml
@@ -16,6 +16,7 @@ jobs:
         with:
           node-version: 16
       - run: npm ci
+      - run: npm run build
       - run: npm test
 
   publish-npm:
