+++ .github/workflows/test.yml
@@ -30,4 +30,5 @@ jobs:
           node-version: ${{ matrix.node-version }}
           cache: npm
       - run: npm ci
+      - run: npm run build
       - run: npm test
